# hjyserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/device/askEd713RealData": {
            "post": {
                "description": "ask Ed713 device to send real data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ED713 device"
                ],
                "summary": "askEd713RealData",
                "parameters": [
                    {
                        "description": "ack Ed713 info",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.AskEd713RealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/askX1RealData": {
            "post": {
                "description": "ask X1 device to send real data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "askX1RealData",
                "parameters": [
                    {
                        "description": "ack X1 info",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.AskX1RealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/cleanX1Event": {
            "post": {
                "description": "clean event from X1 device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "cleanX1Event",
                "parameters": [
                    {
                        "description": "clean X1 event",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.CleanX1EventReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/controlLamp": {
            "post": {
                "description": "control lamp status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "controlLamp",
                "parameters": [
                    {
                        "description": "control lamp",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ControlLampReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/createStudyRoom": {
            "post": {
                "description": "create study room",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "createStudyRoom",
                "parameters": [
                    {
                        "description": "room information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.CreateStudyRoomReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StudyRoom"
                        }
                    }
                }
            }
        },
        "/v1/device/improveDisturbed": {
            "post": {
                "description": "import x1 device disturbed data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "improveDisturbed",
                "parameters": [
                    {
                        "description": "X1 distrube switch",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ImproveDisturbedReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/insert": {
            "post": {
                "description": "insert user device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "insertDevice",
                "parameters": [
                    {
                        "description": "user device information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.NewDeviceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.UserDevice"
                        }
                    }
                }
            }
        },
        "/v1/device/insertFallParams": {
            "post": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "insertFallParams",
                "parameters": [
                    {
                        "description": "fall check device params",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                }
            }
        },
        "/v1/device/modifyStudyRoom": {
            "post": {
                "description": "修改自习室信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "modifyStudyRoom",
                "parameters": [
                    {
                        "description": "room information",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ModifyStudyRoomReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/device/openLampRealData": {
            "post": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "openLampRealData",
                "parameters": [
                    {
                        "description": "lamp real data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.LampRealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/queryAlarmRecord": {
            "get": {
                "description": "query alarm record report from fall check device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryAlarmRecord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallAlarm"
                        }
                    }
                }
            }
        },
        "/v1/device/queryBindByMac": {
            "get": {
                "description": "query device has bind by mac",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryBindByMac",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mdb.DeviceBindResp"
                        }
                    }
                }
            }
        },
        "/v1/device/queryById": {
            "get": {
                "description": "query device infomaion",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryById",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "device id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.Device"
                        }
                    }
                }
            }
        },
        "/v1/device/queryByUser": {
            "get": {
                "description": "query device infomaion by user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryByUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "flag",
                        "name": "flag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.Device"
                        }
                    }
                }
            }
        },
        "/v1/device/queryDateListInReport": {
            "get": {
                "description": "查詢报表中含有数据的日期列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "summary": "queryDateListInReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mdb.QueryDateListResp"
                        }
                    }
                }
            }
        },
        "/v1/device/queryFallCheckStatus": {
            "get": {
                "description": "query fall check data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryFallCheckStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallCheck"
                        }
                    }
                }
            }
        },
        "/v1/device/queryFallExistRecord": {
            "get": {
                "description": "query fall check data if has person exist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryFallExistRecord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallCheck"
                        }
                    }
                }
            }
        },
        "/v1/device/queryFallParams": {
            "get": {
                "description": "query fall check device install params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fall check device"
                ],
                "summary": "queryFallParams",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "device id",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.FallParams"
                        }
                    }
                }
            }
        },
        "/v1/device/queryHeartRate": {
            "get": {
                "description": "query heart data from heart rate table",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sleep device"
                ],
                "summary": "queryHeartRate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day, format yyyy-MM-dd",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day, format yyyy-MM-dd",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.HeartRate"
                        }
                    }
                }
            }
        },
        "/v1/device/queryInviteStudyRoom": {
            "get": {
                "description": "query invited study room by create user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "queryInviteStudyRoom",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StudyRoomUserDetail"
                        }
                    }
                }
            }
        },
        "/v1/device/queryLampControl": {
            "get": {
                "description": "query lamp control data by mac",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "queryLampControl",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.LampControlSql"
                        }
                    }
                }
            }
        },
        "/v1/device/queryLampEvent": {
            "get": {
                "description": "query lamp device real-data by mac and time",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lamp device"
                ],
                "summary": "queryLampEvent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "begin day",
                        "name": "begin_day",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end day",
                        "name": "end_day",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.EventReportSql"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mdb.AskEd713RealDataReq": {
            "type": "object",
            "properties": {
                "freq": {
                    "description": "发送频率\nrequired: false",
                    "type": "integer"
                },
                "keep_push": {
                    "description": "是否保持推送\nrequired: true\nenum: 0,1",
                    "type": "integer"
                },
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
        "mdb.AskX1RealDataReq": {
            "type": "object",
            "properties": {
                "freq": {
                    "description": "发送频率\nrequired: false",
                    "type": "integer"
                },
                "keep_push": {
                    "description": "是否保持推送\nrequired: true\nenum: 0,1",
                    "type": "integer"
                },
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
        "mdb.CleanX1EventReq": {
            "type": "object",
            "properties": {
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "mdb.ImproveDisturbedReq": {
            "type": "object",
            "properties": {
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                },
                "switch": {
                    "description": "开关状态\nrequired: true\nenum: 0,1",
                    "type": "integer"
                }
            }
        },
...
    }
}